refactor(jazz-tools): delete worker-protocol.ts
Of the ~14 typed message-variant interfaces in `worker-protocol.ts`, only
four names were still imported anywhere after the binary-protocol switch
— and three of those were trivially inlinable. The Rust `MainToWorkerWire`
/ `WorkerToMainWire` enums in `worker_protocol.rs` are now the single
source of truth for every envelope except the JS-object `init` message
and the JS-shim's `{type:"ready"}` post.

Inline the survivors and drop the file:

- `WorkerLifecycleEvent` moves into `runtime/worker-bridge.ts` (its only
  two consumers — the bridge adapter and `db.ts::sendLifecycleHint` — both
  already pull from `worker-bridge.ts`).
- `InitMessage` moves into `worker/jazz-worker.ts` (its only consumer is
  the bootstrap shim that parses it before handing the JsValue to
  `runAsWorker`).
- The bridge listener types `LocalBatchRecordsSyncMessage["batches"]` and
  `MutationErrorReplayMessage["batch"]` collapse to `LocalBatchRecord[]`
  and `LocalBatchRecord` directly — they were just one-field wrappers.
- All other interfaces (`SyncToWorkerMessage`, the `Debug*` chain,
  `MainToWorkerMessage` / `WorkerToMainMessage` unions, `Ready` /
  `InitOk` / `UpstreamConnected` / `Shutdown*` / `Error`,
  `SequencedSyncPayload`, etc.) had no remaining consumers — they were
  stale shadows of the binary protocol.

Tests: 1111 vitest pass.
